requirements.txt
aiogram==3.0.0
pyyaml
pydantic==2.6
